+++ plugin-src/button-swap/button-swap/main.c
@@ -6,13 +6,13 @@
 
 #define plugin_name "button-swap"
 
-int32_t (*sceSystemServiceParamGetInt)(int32_t paramId, int32_t *value);
+s32 (*sceSystemServiceParamGetInt)(s32 paramId, s32 *value);
 
 HOOK_INIT(sceSystemServiceParamGetInt);
 
-int32_t sceSystemServiceParamGetInt_hook(int32_t paramId, int32_t *value) {
-    int32_t ret = 0;
-    ret = HOOK_CONTINUE(sceSystemServiceParamGetInt, int32_t(*)(int32_t, int32_t *), paramId, value);
+s32 sceSystemServiceParamGetInt_hook(s32 paramId, s32 *value) {
+    s32 ret = 0;
+    ret = HOOK_CONTINUE(sceSystemServiceParamGetInt, s32(*)(s32, s32 *), paramId, value);
     // swap X and O around in Asia PS4
     if (paramId == 1000) {
         *value = !*value;
@@ -20,17 +20,17 @@ int32_t sceSystemServiceParamGetInt_hook(int32_t paramId, int32_t *value) {
     return ret;
 }
 
-int attr_module_hidden module_start(size_t argc, const void *args) {
+s32 attr_module_hidden module_start(s64 argc, const void *args) {
     final_printf("[GoldHEN] <%s> %s\n", plugin_name, __func__);
     boot_ver();
-    int h = 0;
+    s32 h = 0;
     sys_dynlib_load_prx("libSceSystemService.sprx", &h);
     sys_dynlib_dlsym(h, "sceSystemServiceParamGetInt", &sceSystemServiceParamGetInt);
     HOOK32(sceSystemServiceParamGetInt);
     return 0;
 }
 
-int attr_module_hidden module_stop(size_t argc, const void *args) {
+s32 attr_module_hidden module_stop(s64 argc, const void *args) {
     final_printf("[GoldHEN] <%s> %s\n", plugin_name, __func__);
     UNHOOK(sceSystemServiceParamGetInt);
     return 0;
