+++ plugin-src/fliprate-remover/fliprate-remover/main.c
@@ -13,19 +13,19 @@
 
 HOOK_INIT(sceVideoOutSetFlipRate);
 
-s32 sceVideoOutSetFlipRate_hook(int handle, int fliprate){
+s32 sceVideoOutSetFlipRate_hook(s32 handle, s32 fliprate){
     debug_printf("handle 0x%08x fliprate 0x%08x\n", handle, fliprate);
     return 0;
 }
 
-s32 attr_module_hidden module_start(size_t argc, const void *args) {
+s32 attr_module_hidden module_start(s64 argc, const void *args) {
     final_printf("[GoldHEN] <%s> %s\n", plugin_name, __func__);
     boot_ver();
     HOOK32(sceVideoOutSetFlipRate);
     return 0;
 }
 
-s32 attr_module_hidden module_stop(size_t argc, const void *args) {
+s32 attr_module_hidden module_stop(s64 argc, const void *args) {
     final_printf("[GoldHEN] <%s> %s\n", plugin_name, __func__);
     UNHOOK(sceVideoOutSetFlipRate);
     return 0;
